requirements.txt
# Python compatibility: 3.11 or 3.12 (Windows-friendly pinned set)

# Core ML stack
torch==2.5.1
lightning==2.6.1
pytorch-lightning==2.6.1
pytorch-forecasting==1.6.1
stable-baselines3==2.5.0
gymnasium==0.29.1
xgboost==2.1.4

# Data & numerics
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.5.2
joblib==1.4.2
ta==0.11.0
pyarrow==18.1.0

# Exchange & networking
kucoin-python==1.0.11
websockets==12.0
httpx==0.27.2
tenacity==8.5.0
requests==2.32.3

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Cache / queue
redis==5.2.1

# API & dashboard
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
plotly==5.24.1

# Security & config
cryptography==44.0.0
python-dotenv==1.0.1
pydantic==2.10.4

# Utilities
loguru==0.7.3
schedule==1.2.2
google-generativeai==0.8.6

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
